refactor: 선택된 방 store 분리 #19
- selectedRoom 파일을 store와 type으로 분리
