+++ app/prayer/creation.tsx
@@ -2,7 +2,7 @@ import React, { useState, useMemo, useEffect, useRef } from "react";
 import { StyleSheet, SafeAreaView, BackHandler } from "react-native";
 import { color } from "../../src/common/styles/color";
 import { useRoomMembersQuery } from "../../src/domain/prayerRoom/hooks/queries/roomQueries";
-import { useSelectedRoomStore } from "../../src/domain/prayerRoom/types/selectedRoomStore";
+import { useSelectedRoomStore } from "../../src/domain/prayerRoom/stores/useSelectedRoomStore";
 import { RoomMember } from "../../src/domain/prayerRoom/types/dto/response/roomMember";
 import { router } from "expo-router";
 import PrayerMemberSelectionModal from "../../src/domain/PrayerCreation/components/modal/PrayerMemberSelectionModal";
