+++ src/domain/PrayerCreation/components/PrayerCreationBody.tsx
@@ -7,7 +7,7 @@ import { SelectedMember } from "../types/SelectedMember";
 import { KeyboardAwareScrollView } from "react-native-keyboard-aware-scroll-view";
 import { usePrayerCreationStore } from "./../stores/usePrayerCreationStore";
 import { PrayerCreationItem } from "../types/PrayerCreationItem";
-import { useSelectedRoomStore } from "../../prayerRoom/types/selectedRoomStore";
+import { useSelectedRoomStore } from "../../prayerRoom/stores/useSelectedRoomStore";
 import { useRoomMembersQuery } from "../../prayerRoom/hooks/queries/roomQueries";
 import { RoomMember } from "../../prayerRoom/types/dto/response/roomMember";
