+++ src/domain/prayerRoom/stores/useSelectedRoomStore.ts
@@ -1,16 +1,6 @@
 import { create } from "zustand";
+import { SelectedRoomStore, SelectedRoom } from "../types/selectedRoom";
 import { Room } from "../../roomList/types/dto/responses/room";
-
-export interface SelectedRoom {
-  selectedRoom: Room | null;
-}
-
-export interface SelectedRoomAction {
-  selectRoom: (room: Room) => void;
-  resetRoom: () => void;
-}
-
-export type SelectedRoomStore = SelectedRoom & SelectedRoomAction;
 
 const initialState: SelectedRoom = {
   selectedRoom: null,
